fix(review): apply safe-auto code review fixes
- fix(wizard): rename ManualBankTransfer → BankTransfer in Step4Payment
  (was failing Zod validation and backend enum parse)
- fix(tds): remove double-count of surcharge+cess in annualTaxLiability
  (AnnualProjectedTax already includes both)
- fix(pt): use payroll period date for PT slab lookup instead of DateTime.UtcNow
- fix(pt): move HalfYearlySplit salary filter before OrderByDescending
- fix(tds): correct Section 87A rebate — check only income ceiling, not annualTax <= rebateAmount
- fix(bank-advice): remove (double) cast on monetary decimal in XLS generator
- chore(repo): remove committed TODO comment from WorkLocationRepository
- fix(types): dateOfBirth nullable in EmployeeDto TypeScript type

+++ src/Payroll.Infrastructure/Services/BankAdviceGenerator.cs
@@ -36,7 +36,7 @@ public sealed class BankAdviceGenerator : IBankAdviceGenerator
         {
             ws.Cell(row, 1).Value = item.EmployeeCode;
             ws.Cell(row, 2).Value = item.EmployeeName;
-            ws.Cell(row, 3).Value = (double)item.Amount;
+            ws.Cell(row, 3).Value = item.Amount;
             ws.Cell(row, 3).Style.NumberFormat.Format = "#,##0.00";
             ws.Cell(row, 4).Value = item.BankName;
             ws.Cell(row, 5).Value = item.BankAccountNo;
